examples/discord-welcomer/requirements.txt
pincer
python-dotenv
aiohttp
